requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-CORS==3.0.10
requests==2.26.0
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==0.19.2
firebase-admin==5.2.0
click==8.1.3
